Move BH curve fitting function to preprocessing utilities
